+++ pkg/manager/impl/execution_manager.go
@@ -184,39 +184,6 @@ func (m *ExecutionManager) addPluginOverrides(ctx context.Context, executionID *
 	return nil, nil
 }
 
-type completeTaskResources struct {
-	Defaults runtimeInterfaces.TaskResourceSet
-	Limits   runtimeInterfaces.TaskResourceSet
-}
-
-func getTaskResourcesAsSet(ctx context.Context, identifier *core.Identifier,
-	resourceEntries []*core.Resources_ResourceEntry, resourceName string) runtimeInterfaces.TaskResourceSet {
-
-	result := runtimeInterfaces.TaskResourceSet{}
-	for _, entry := range resourceEntries {
-		switch entry.Name {
-		case core.Resources_CPU:
-			result.CPU = util.ParseQuantityNoError(ctx, identifier.String(), fmt.Sprintf("%v.cpu", resourceName), entry.Value)
-		case core.Resources_MEMORY:
-			result.Memory = util.ParseQuantityNoError(ctx, identifier.String(), fmt.Sprintf("%v.memory", resourceName), entry.Value)
-		case core.Resources_EPHEMERAL_STORAGE:
-			result.EphemeralStorage = util.ParseQuantityNoError(ctx, identifier.String(),
-				fmt.Sprintf("%v.ephemeral storage", resourceName), entry.Value)
-		case core.Resources_GPU:
-			result.GPU = util.ParseQuantityNoError(ctx, identifier.String(), "gpu", entry.Value)
-		}
-	}
-
-	return result
-}
-
-func getCompleteTaskResourceRequirements(ctx context.Context, identifier *core.Identifier, task *core.CompiledTask) completeTaskResources {
-	return completeTaskResources{
-		Defaults: getTaskResourcesAsSet(ctx, identifier, task.GetTemplate().GetContainer().Resources.Requests, "requests"),
-		Limits:   getTaskResourcesAsSet(ctx, identifier, task.GetTemplate().GetContainer().Resources.Limits, "limits"),
-	}
-}
-
 // TODO: Delete this code usage after the flyte v0.17.0 release
 // Assumes input contains a compiled task with a valid container resource execConfig.
 //
@@ -252,7 +219,7 @@ func (m *ExecutionManager) setCompiledTaskDefaults(ctx context.Context, task *co
 	// The IDL representation for container-type tasks represents resources as a list with string quantities.
 	// In order to easily reason about them we convert them to a set where we can O(1) fetch specific resources (e.g. CPU)
 	// and represent them as comparable quantities rather than strings.
-	taskResourceRequirements := getCompleteTaskResourceRequirements(ctx, task.Template.Id, task)
+	taskResourceRequirements := util.GetCompleteTaskResourceRequirements(ctx, task.Template.Id, task)
 
 	cpu := flytek8s.AdjustOrDefaultResource(taskResourceRequirements.Defaults.CPU, taskResourceRequirements.Limits.CPU,
 		platformTaskResources.Defaults.CPU, platformTaskResources.Limits.CPU)
@@ -331,68 +298,6 @@ func (m *ExecutionManager) setCompiledTaskDefaults(ctx context.Context, task *co
 		Limits:   finalizedResourceLimits,
 	}
 }
-
-/*func parseQuantityNoError(ctx context.Context, ownerID, name, value string) resource.Quantity {
-	q, err := resource.ParseQuantity(value)
-	if err != nil {
-		logger.Infof(ctx, "Failed to parse owner's [%s] resource [%s]'s value [%s] with err: %v", ownerID, name, value, err)
-	}
-
-	return q
-}
-
-func fromAdminProtoTaskResourceSpec(ctx context.Context, spec *admin.TaskResourceSpec) runtimeInterfaces.TaskResourceSet {
-	result := runtimeInterfaces.TaskResourceSet{}
-	if len(spec.Cpu) > 0 {
-		result.CPU = parseQuantityNoError(ctx, "project", "cpu", spec.Cpu)
-	}
-
-	if len(spec.Memory) > 0 {
-		result.Memory = parseQuantityNoError(ctx, "project", "memory", spec.Memory)
-	}
-
-	if len(spec.Storage) > 0 {
-		result.Storage = parseQuantityNoError(ctx, "project", "storage", spec.Storage)
-	}
-
-	if len(spec.EphemeralStorage) > 0 {
-		result.EphemeralStorage = parseQuantityNoError(ctx, "project", "ephemeral storage", spec.EphemeralStorage)
-	}
-
-	if len(spec.Gpu) > 0 {
-		result.GPU = parseQuantityNoError(ctx, "project", "gpu", spec.Gpu)
-	}
-
-	return result
-}*/
-
-/*func (m *ExecutionManager) getTaskResources(ctx context.Context, workflow *core.Identifier) workflowengineInterfaces.TaskResources {
-	resource, err := m.resourceManager.GetResource(ctx, interfaces.ResourceRequest{
-		Project:      workflow.Project,
-		Domain:       workflow.Domain,
-		Workflow:     workflow.Name,
-		ResourceType: admin.MatchableResource_TASK_RESOURCE,
-	})
-
-	if err != nil {
-		logger.Warningf(ctx, "Failed to fetch override values when assigning task resource default values for [%+v]: %v",
-			workflow, err)
-	}
-
-	logger.Debugf(ctx, "Assigning task requested resources for [%+v]", workflow)
-	var taskResourceAttributes = workflowengineInterfaces.TaskResources{}
-	if resource != nil && resource.Attributes != nil && resource.Attributes.GetTaskResourceAttributes() != nil {
-		taskResourceAttributes.Defaults = fromAdminProtoTaskResourceSpec(ctx, resource.Attributes.GetTaskResourceAttributes().Defaults)
-		taskResourceAttributes.Limits = fromAdminProtoTaskResourceSpec(ctx, resource.Attributes.GetTaskResourceAttributes().Limits)
-	} else {
-		taskResourceAttributes = workflowengineInterfaces.TaskResources{
-			Defaults: m.config.TaskResourceConfiguration().GetDefaults(),
-			Limits:   m.config.TaskResourceConfiguration().GetLimits(),
-		}
-	}
-
-	return taskResourceAttributes
-}*/
 
 // Fetches inherited execution metadata including the parent node execution db model id and the source execution model id
 // as well as sets request spec metadata with the inherited principal and adjusted nesting data.
