+++ pkg/manager/impl/validation/task_validator_test.go
@@ -6,28 +6,28 @@ import (
 	"errors"
 	"testing"
 
-	"google.golang.org/protobuf/types/known/structpb"
-
-	corev1 "k8s.io/api/core/v1"
-
-	"github.com/flyteorg/flyteidl/gen/pb-go/flyteidl/core"
-	"k8s.io/apimachinery/pkg/api/resource"
-
 	"github.com/flyteorg/flyteadmin/pkg/manager/impl/testutils"
 	runtimeInterfaces "github.com/flyteorg/flyteadmin/pkg/runtime/interfaces"
 	runtimeMocks "github.com/flyteorg/flyteadmin/pkg/runtime/mocks"
+	workflowengineInterfaces "github.com/flyteorg/flyteadmin/pkg/workflowengine/interfaces"
+	"github.com/flyteorg/flyteidl/gen/pb-go/flyteidl/core"
+
 	"github.com/stretchr/testify/assert"
+
+	"google.golang.org/protobuf/types/known/structpb"
+
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/resource"
 )
 
-func getMockTaskConfigProvider() runtimeInterfaces.TaskResourceConfiguration {
-	var taskConfig = runtimeMocks.MockTaskResourceConfiguration{}
-	taskConfig.Limits = runtimeInterfaces.TaskResourceSet{
-		Memory: resource.MustParse("500Mi"),
-		CPU:    resource.MustParse("200m"),
-		GPU:    resource.MustParse("8"),
+func getMockTaskResources() workflowengineInterfaces.TaskResources {
+	return workflowengineInterfaces.TaskResources{
+		Limits: runtimeInterfaces.TaskResourceSet{
+			Memory: resource.MustParse("500Mi"),
+			CPU:    resource.MustParse("200m"),
+			GPU:    resource.MustParse("8"),
+		}
 	}
-
-	return &taskConfig
 }
 
 var mockWhitelistConfigProvider = runtimeMocks.NewMockWhitelistConfiguration()
@@ -47,26 +47,26 @@ func TestValidateTask(t *testing.T) {
 	}
 	request.Spec.Template.GetContainer().Resources = &core.Resources{Requests: resources}
 	err := ValidateTask(context.Background(), request, testutils.GetRepoWithDefaultProject(),
-		getMockTaskConfigProvider(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
+		getMockTaskResources(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
 	assert.EqualError(t, err, "Requested CPU default [1536Mi] is greater than current limit set in the platform configuration [200m]. Please contact Flyte Admins to change these limits or consult the configuration")
 
 	request.Spec.Template.Target = &core.TaskTemplate_K8SPod{K8SPod: &core.K8SPod{}}
 	err = ValidateTask(context.Background(), request, testutils.GetRepoWithDefaultProject(),
-		getMockTaskConfigProvider(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
+		getMockTaskResources(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
 	assert.EqualError(t, err, "invalid TaskSpecification, pod tasks should specify their target as a K8sPod with a defined pod spec")
 
 	resourceList := corev1.ResourceList{corev1.ResourceCPU: resource.MustParse("1.5Gi")}
 	podSpec := &corev1.PodSpec{Containers: []corev1.Container{{Resources: corev1.ResourceRequirements{Requests: resourceList}}}}
 	request.Spec.Template.Target = &core.TaskTemplate_K8SPod{K8SPod: &core.K8SPod{PodSpec: transformStructToStructPB(t, podSpec)}}
 	err = ValidateTask(context.Background(), request, testutils.GetRepoWithDefaultProject(),
-		getMockTaskConfigProvider(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
+		getMockTaskResources(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
 	assert.EqualError(t, err, "Requested CPU default [1536Mi] is greater than current limit set in the platform configuration [200m]. Please contact Flyte Admins to change these limits or consult the configuration")
 
 	resourceList = corev1.ResourceList{corev1.ResourceCPU: resource.MustParse("200m")}
 	podSpec = &corev1.PodSpec{Containers: []corev1.Container{{Resources: corev1.ResourceRequirements{Requests: resourceList}}}}
 	request.Spec.Template.Target = &core.TaskTemplate_K8SPod{K8SPod: &core.K8SPod{PodSpec: transformStructToStructPB(t, podSpec)}}
 	err = ValidateTask(context.Background(), request, testutils.GetRepoWithDefaultProject(),
-		getMockTaskConfigProvider(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
+		getMockTaskResources(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
 	assert.Nil(t, err)
 }
 
@@ -85,14 +85,14 @@ func TestValidateTaskEmptyProject(t *testing.T) {
 	request := testutils.GetValidTaskRequest()
 	request.Id.Project = ""
 	err := ValidateTask(context.Background(), request, testutils.GetRepoWithDefaultProject(),
-		getMockTaskConfigProvider(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
+		getMockTaskResources(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
 	assert.EqualError(t, err, "missing project")
 }
 
 func TestValidateTaskInvalidProjectAndDomain(t *testing.T) {
 	request := testutils.GetValidTaskRequest()
 	err := ValidateTask(context.Background(), request, testutils.GetRepoWithDefaultProjectAndErr(errors.New("foo")),
-		getMockTaskConfigProvider(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
+		getMockTaskResources(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
 	assert.EqualError(t, err, "failed to validate that project [project] and domain [domain] are registered, err: [foo]")
 }
 
@@ -100,7 +100,7 @@ func TestValidateTaskEmptyDomain(t *testing.T) {
 	request := testutils.GetValidTaskRequest()
 	request.Id.Domain = ""
 	err := ValidateTask(context.Background(), request, testutils.GetRepoWithDefaultProject(),
-		getMockTaskConfigProvider(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
+		getMockTaskResources(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
 	assert.EqualError(t, err, "missing domain")
 }
 
@@ -108,7 +108,7 @@ func TestValidateTaskEmptyName(t *testing.T) {
 	request := testutils.GetValidTaskRequest()
 	request.Id.Name = ""
 	err := ValidateTask(context.Background(), request, testutils.GetRepoWithDefaultProject(),
-		getMockTaskConfigProvider(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
+		getMockTaskResources(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
 	assert.EqualError(t, err, "missing name")
 }
 
@@ -116,7 +116,7 @@ func TestValidateTaskEmptyVersion(t *testing.T) {
 	request := testutils.GetValidTaskRequest()
 	request.Id.Version = ""
 	err := ValidateTask(context.Background(), request, testutils.GetRepoWithDefaultProject(),
-		getMockTaskConfigProvider(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
+		getMockTaskResources(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
 	assert.EqualError(t, err, "missing version")
 }
 
@@ -124,7 +124,7 @@ func TestValidateTaskEmptyType(t *testing.T) {
 	request := testutils.GetValidTaskRequest()
 	request.Spec.Template.Type = ""
 	err := ValidateTask(context.Background(), request, testutils.GetRepoWithDefaultProject(),
-		getMockTaskConfigProvider(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
+		getMockTaskResources(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
 	assert.EqualError(t, err, "missing type")
 }
 
@@ -132,7 +132,7 @@ func TestValidateTaskEmptyMetadata(t *testing.T) {
 	request := testutils.GetValidTaskRequest()
 	request.Spec.Template.Metadata = nil
 	err := ValidateTask(context.Background(), request, testutils.GetRepoWithDefaultProject(),
-		getMockTaskConfigProvider(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
+		getMockTaskResources(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
 	assert.EqualError(t, err, "missing metadata")
 }
 
@@ -140,7 +140,7 @@ func TestValidateTaskEmptyRuntimeVersion(t *testing.T) {
 	request := testutils.GetValidTaskRequest()
 	request.Spec.Template.Metadata.Runtime.Version = ""
 	err := ValidateTask(context.Background(), request, testutils.GetRepoWithDefaultProject(),
-		getMockTaskConfigProvider(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
+		getMockTaskResources(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
 	assert.EqualError(t, err, "missing runtime version")
 }
 
@@ -148,7 +148,7 @@ func TestValidateTaskEmptyTypedInterface(t *testing.T) {
 	request := testutils.GetValidTaskRequest()
 	request.Spec.Template.Interface = nil
 	err := ValidateTask(context.Background(), request, testutils.GetRepoWithDefaultProject(),
-		getMockTaskConfigProvider(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
+		getMockTaskResources(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
 	assert.EqualError(t, err, "missing typed interface")
 }
 
@@ -156,7 +156,7 @@ func TestValidateTaskEmptyContainer(t *testing.T) {
 	request := testutils.GetValidTaskRequest()
 	request.Spec.Template.Target = nil
 	err := ValidateTask(context.Background(), request, testutils.GetRepoWithDefaultProject(),
-		getMockTaskConfigProvider(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
+		getMockTaskResources(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
 	assert.Nil(t, err)
 }
 
@@ -164,7 +164,7 @@ func TestValidateTaskEmptyImage(t *testing.T) {
 	request := testutils.GetValidTaskRequest()
 	request.Spec.Template.GetContainer().Image = ""
 	err := ValidateTask(context.Background(), request, testutils.GetRepoWithDefaultProject(),
-		getMockTaskConfigProvider(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
+		getMockTaskResources(), mockWhitelistConfigProvider, taskApplicationConfigProvider)
 	assert.EqualError(t, err, "missing image")
 }
